Remove macos dependent emulator code
The reasoning behind this is that the emulator should exist in $PATH already - react's own getting started guide asks you to set this up. It should therefore be expected that the developer has $PATH set up such that `emulator` can be reached from CLI.

Also, platform-specific code is a huge issue.